binder: fix include/nuttx/android/binder.h:167:18: error: conflicting declaration ‘typedef uint64_t binder_size_t’
  167 | typedef uint64_t binder_size_t;
      |                  ^~~~~~~~~~~~~
In file included from /home/xiaoxiang/vela/apps/external/android/binder/binder/include/binder/IPCThreadState.h:20,
                 from binder/Parcel.cpp:36:
/home/xiaoxiang/vela/apps/external/android/binder/binder/include/binder/Parcel.h:41:28: note: previous declaration as ‘typedef long long unsigned int binder_size_t’
   41 | typedef unsigned long long binder_size_t;

+++ libs/binder/include/binder/Parcel.h
@@ -35,10 +35,10 @@
 
 #ifdef BINDER_IPC_32BIT
 //NOLINTNEXTLINE(google-runtime-int) b/173188702
-typedef unsigned int binder_size_t;
+typedef uint32_t binder_size_t;
 #else
 //NOLINTNEXTLINE(google-runtime-int) b/173188702
-typedef unsigned long long binder_size_t;
+typedef uint64_t binder_size_t;
 #endif
 
 struct flat_binder_object;
